Move PDF test utils to tests.pdf_utils
